sample-plugin.go: Support session per node
This patch changes the way sessions are handled.
Session will be created and reused.
There are two ways of using sessions, one the default session, that can
reach any node in the cluster.
The other, when queryHost is set, use a session that points only to that
host.

This is useful when querying local table (like those under the system
keyspace).

The sessions are saved in a map where the host ip is the key, the empty
string is used for the default connection.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>