datasource.go: All hosts support
This patch handles two issues around multiple hosts.
First it support the use of a query parameter: allHosts that set how the
plugin would support the queryHost parameter, if it is set to true data
will be read from all hosts, this is useful for system tables.
If not, one host will be connected to.

Second, to overcome a bug with the golang driver, for reading from a
specific host a new cluster object will be created for each connection.

Lastly a gofmt was run on the file